Removing unnecessary elements from project
Removing css & less files as project will use scss.

Adjusted ping view to ensure previously mentioned routes are displayed which are not related to vue.js.

Removed django_cron as it is not needed for this project.

Removed vuetify as we will be using custom components and styling with TailwindCSS.

Updated readme.md to let users know to create migrations first (for django_cron, but still it's better to mention it).

Imported tailwindcss modules into scss file.